+++ Lodgify/Models/Person.cs
@@ -7,11 +7,11 @@ namespace Lodgify.Models
 {
     public class Person
     {
-        private long id;
+        private int id;
         private string firstName;
         private string lastName;
 
-        public long Id { get { return id; } set { id = value; } }
+        public int Id { get { return id; } set { id = value; } }
         public string FirstName { get { return firstName; } set { firstName = value; } }
         public string LastName { get { return lastName; } set { lastName = value; } }
